Extract Elasticsearch types and helpers into esutil package
