+++ go/v1beta1/storage/elasticsearch.go
@@ -19,13 +19,13 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"io"
 
 	"github.com/elastic/go-elasticsearch/v7"
 	"github.com/golang/protobuf/proto"
 	"github.com/golang/protobuf/ptypes"
 	"github.com/google/uuid"
 	"github.com/rode/grafeas-elasticsearch/go/config"
+	. "github.com/rode/grafeas-elasticsearch/go/v1beta1/storage/esutil"
 	"github.com/rode/grafeas-elasticsearch/go/v1beta1/storage/filtering"
 	"github.com/rode/grafeas-elasticsearch/go/v1beta1/storage/migration"
 	"go.uber.org/zap"
@@ -364,7 +364,7 @@ func (es *ElasticsearchStorage) BatchCreateOccurrences(ctx context.Context, proj
 	}
 
 	response := &esBulkResponse{}
-	err = decodeResponse(res.Body, response)
+	err = DecodeResponse(res.Body, response)
 	if err != nil {
 		return nil, []error{
 			createError(log, "error decoding ES response", nil),
@@ -596,7 +596,7 @@ func (es *ElasticsearchStorage) BatchCreateNotes(ctx context.Context, projectId,
 	}
 
 	searchResponse := &esMultiSearchResponse{}
-	err = decodeResponse(res.Body, searchResponse)
+	err = DecodeResponse(res.Body, searchResponse)
 	if err != nil {
 		return nil, []error{
 			createError(log, "error decoding ES response", nil),
@@ -657,7 +657,7 @@ func (es *ElasticsearchStorage) BatchCreateNotes(ctx context.Context, projectId,
 	}
 
 	bulkResponse := &esBulkResponse{}
-	err = decodeResponse(res.Body, bulkResponse)
+	err = DecodeResponse(res.Body, bulkResponse)
 	if err != nil {
 		return nil, append(errs, createError(log, "error decoding ES response", nil))
 	}
@@ -726,7 +726,7 @@ func (es *ElasticsearchStorage) GetVulnerabilityOccurrencesSummary(ctx context.C
 }
 
 func (es *ElasticsearchStorage) genericGet(ctx context.Context, log *zap.Logger, search *esSearch, index string, protoMessage interface{}) (string, error) {
-	encodedBody, requestJson := encodeRequest(search)
+	encodedBody, requestJson := EncodeRequest(search)
 	log = log.With(zap.String("request", requestJson))
 
 	res, err := es.client.Search(
@@ -742,7 +742,7 @@ func (es *ElasticsearchStorage) genericGet(ctx context.Context, log *zap.Logger,
 	}
 
 	var searchResults esSearchResponse
-	if err := decodeResponse(res.Body, &searchResults); err != nil {
+	if err := DecodeResponse(res.Body, &searchResults); err != nil {
 		return "", createError(log, "error unmarshalling elasticsearch response", err)
 	}
 
@@ -775,7 +775,7 @@ func (es *ElasticsearchStorage) genericCreate(ctx context.Context, log *zap.Logg
 	}
 
 	esResponse := &esIndexDocResponse{}
-	err = decodeResponse(res.Body, esResponse)
+	err = DecodeResponse(res.Body, esResponse)
 	if err != nil {
 		return createError(log, "error decoding elasticsearch response", err)
 	}
@@ -807,7 +807,7 @@ func (es *ElasticsearchStorage) genericUpdate(ctx context.Context, log *zap.Logg
 	}
 
 	esResponse := &esIndexDocResponse{}
-	err = decodeResponse(res.Body, esResponse)
+	err = DecodeResponse(res.Body, esResponse)
 	if err != nil {
 		return createError(log, "error decoding elasticsearch response", err)
 	}
@@ -818,7 +818,7 @@ func (es *ElasticsearchStorage) genericUpdate(ctx context.Context, log *zap.Logg
 }
 
 func (es *ElasticsearchStorage) genericDelete(ctx context.Context, log *zap.Logger, search *esSearch, index string) error {
-	encodedBody, requestJson := encodeRequest(search)
+	encodedBody, requestJson := EncodeRequest(search)
 	log = log.With(zap.String("request", requestJson))
 
 	res, err := es.client.DeleteByQuery(
@@ -835,7 +835,7 @@ func (es *ElasticsearchStorage) genericDelete(ctx context.Context, log *zap.Logg
 	}
 
 	var deletedResults esDeleteResponse
-	if err = decodeResponse(res.Body, &deletedResults); err != nil {
+	if err = DecodeResponse(res.Body, &deletedResults); err != nil {
 		return createError(log, "error unmarshalling elasticsearch response", err)
 	}
 
@@ -864,7 +864,7 @@ func (es *ElasticsearchStorage) genericList(ctx context.Context, log *zap.Logger
 		}
 	}
 
-	encodedBody, requestJson := encodeRequest(body)
+	encodedBody, requestJson := EncodeRequest(body)
 	log = log.With(zap.String("request", requestJson))
 	log.Debug("performing search")
 
@@ -882,7 +882,7 @@ func (es *ElasticsearchStorage) genericList(ctx context.Context, log *zap.Logger
 	}
 
 	var searchResults esSearchResponse
-	if err := decodeResponse(res.Body, &searchResults); err != nil {
+	if err := DecodeResponse(res.Body, &searchResults); err != nil {
 		return nil, createError(log, "error decoding elasticsearch response", err)
 	}
 
@@ -898,20 +898,6 @@ func createError(log *zap.Logger, message string, err error, fields ...zap.Field
 
 	log.Error(message, append(fields, zap.Error(err))...)
 	return status.Errorf(codes.Internal, "%s: %s", message, err)
-}
-
-func decodeResponse(r io.ReadCloser, i interface{}) error {
-	return json.NewDecoder(r).Decode(i)
-}
-
-func encodeRequest(body interface{}) (io.Reader, string) {
-	b, err := json.Marshal(body)
-	if err != nil {
-		// we should know that `body` is a serializable struct before invoking `encodeRequest`
-		panic(err)
-	}
-
-	return bytes.NewReader(b), string(b)
 }
 
 // DeleteByQuery does not support `wait_for` value, although API docs say it is available.
