+++ go/v1beta1/storage/migration/index.go
@@ -26,6 +26,7 @@ import (
 	"strings"
 
 	"github.com/elastic/go-elasticsearch/v7"
+	. "github.com/rode/grafeas-elasticsearch/go/v1beta1/storage/esutil"
 	"go.uber.org/zap"
 )
 
@@ -37,49 +38,6 @@ const (
 	IndexPrefix = "grafeas"
 	AliasPrefix = "grafeas"
 )
-
-type IndexManager interface {
-	LoadMappings(mappingsDir string) error
-	CreateIndex(ctx context.Context, info *IndexInfo, checkExists bool) error
-
-	ProjectsIndex() string
-	ProjectsAlias() string
-
-	OccurrencesIndex(projectId string) string
-	OccurrencesAlias(projectId string) string
-
-	NotesIndex(projectId string) string
-	NotesAlias(projectId string) string
-
-	IncrementIndexVersion(indexName string) string
-	GetLatestVersionForDocumentKind(documentKind string) string
-	GetAliasForIndex(indexName string) string
-}
-
-type VersionedMapping struct {
-	Version  string                 `json:"version"`
-	Mappings map[string]interface{} `json:"mappings"`
-}
-
-type IndexInfo struct {
-	Index        string
-	Alias        string
-	DocumentKind string
-}
-
-type IndexNameParts struct {
-	DocumentKind string
-	Version      string
-	ProjectId    string
-}
-
-type EsIndexManager struct {
-	logger            *zap.Logger
-	client            *elasticsearch.Client
-	projectMapping    *VersionedMapping
-	occurrenceMapping *VersionedMapping
-	noteMapping       *VersionedMapping
-}
 
 func NewEsIndexManager(logger *zap.Logger, client *elasticsearch.Client) *EsIndexManager {
 	return &EsIndexManager{
@@ -166,7 +124,7 @@ func (em *EsIndexManager) CreateIndex(ctx context.Context, info *IndexInfo, chec
 		}
 	}
 
-	payload := encodeRequest(&createIndexReq)
+	payload, _ := EncodeRequest(&createIndexReq)
 	res, err := em.client.Indices.Create(info.Index, em.client.Indices.Create.WithContext(ctx), em.client.Indices.Create.WithBody(payload))
 	if err != nil {
 		return fmt.Errorf("error creating index %s: %s", info.Index, err)
@@ -175,7 +133,9 @@ func (em *EsIndexManager) CreateIndex(ctx context.Context, info *IndexInfo, chec
 	if res.IsError() {
 		if res.StatusCode == http.StatusBadRequest {
 			errResponse := ESErrorResponse{}
-			_ = decodeResponse(res.Body, &errResponse)
+			if err := DecodeResponse(res.Body, &errResponse); err != nil {
+				return fmt.Errorf("error decoding Elasticsearch error response: %s", err)
+			}
 
 			if errResponse.Error.Type == "resource_already_exists_exception" {
 				log.Info("index already exists")
